requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
numpy==1.18.0
Pillow==7.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pyrsistent==0.15.6
pytz==2019.3
PyYAML==5.2
reportlab==3.5.32
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.8
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wincertstore==0.2
zipp==0.6.0
